implemented layout user admin in project

+++ app/Http/Controllers/Admin/Permission/UserController.php
@@ -4,6 +4,7 @@ namespace App\Http\Controllers\Admin\Permission;
 
 use App\Http\Controllers\Controller;
 use App\Http\Requests\Admin\UserAdmin\DestroyRequest;
+use App\Http\Requests\Admin\UserAdmin\SetPermissionRequest;
 use App\Models\User;
 use Illuminate\Auth\Access\AuthorizationException;
 use Illuminate\Contracts\View\View;
@@ -42,9 +43,39 @@ class UserController extends Controller
      */
     public function show(User $user): View
     {
-        $this->authorize("see-admin");
+        $this->authorize("see-user");
 
         return view("admin.user_admin.show", compact("user"));
+    }
+
+    /**
+     * @throws AuthorizationException
+     */
+    public function permissions(User $user): View
+    {
+
+        $this->authorize("allow-to-set-permission");
+
+        $title = "ثبت دسترسی برای کاربر ادمین";
+
+        return view("admin.user_admin.user_permission", compact("user", "title"));
+
+    }
+
+    /**
+     * @param User $user
+     * @param SetPermissionRequest $request
+     * @return RedirectResponse
+     */
+    public function setPermissions(User $user, SetPermissionRequest $request): RedirectResponse
+    {
+
+        $user->permissions()->sync($request->input("permissions"));
+
+        $user->roles()->sync($request->input("roles"));
+
+        return redirect(route("admin.users.index"))->with("success", "دسترسی و مقام های مورد نظر با موفقیت برای کاربر مورد نظر ثبت شد");
+
     }
 
     /**
@@ -53,9 +84,14 @@ class UserController extends Controller
      */
     public function destroy(DestroyRequest $request): RedirectResponse
     {
-        $user = User::find($request->input("user_id"));
 
-        $user?->delete();
+        if ($request->user()->id != $request->input("user_id")) {
+
+            $user = User::find($request->input("user_id"));
+
+            $user?->delete();
+
+        }
 
         return redirect(route("admin.users.index"))->with("success", "کاربر ادمین با موفقیت حذف شد .");
     }
